Autogen experiment 2
3 autogen shots (way more shots within the autogen shots of course)

PROMPT 1:
<previous_prompt>
I have an incomplete project with a jumble of 'Vue2 Component's and 'Vue2.7+ SFC's in the components/ directory of the project. Some of the Vue2.7+ SFCs have .js extensions but they are not .js files, treat them as .vue files. I want you to turn this project into a proper Vue 3 project for me including setting up all of the npm dependencies, App.vue file, README.md, and scripts to build and serve the Vue 3 application. This means that if files are missing then you should create them. If files need to be updated or corrected, do so. If you need to convert Vue 2 components into Vue 3 SFCs, do so. The Vue 3 application should make use of all of the Vue 3 components. When you think you are done, re-evaluate the project again by listing and reading the files to ensure that everything looks good. If not generate a new plan on how to fix those issues that you see.
</previous_prompt>

PROMPT 2 added on:
I have supplied the previous prompt above since you were in the middle of solving that task and you finished converting the components to .vue files but didn't finish the rest of the task. Please look over the previous prompt and the current state of the project and finish the task.

PROMPT 3 added on:
The last thing you failed on was creating the package.json and also I noticed that the App.vue doesn't use any of the components/*.vue files. Please make sure that you do.