autogen experiment 4
Prompt:
I have an incomplete project with a jumble of 'Vue2 Component's and 'Vue2.7+ SFC's in the components/ directory of the project. Some of the Vue2.7+ SFCs have .js extensions but they are not .js files, treat them as .vue files. I want you to turn this project into a proper Vue 3 project for me including setting up all of the npm dependencies (package.json), App.vue file, README.md, index.html, and scripts to build and serve the Vue 3 application. This means that if files are missing then you should create them. If files need to be updated or corrected, do so. If you need to convert Vue 2 components into Vue 3 SFCs, do so. The Vue 3 application should make use of all of the Vue 3 components (components/*.vue files). When you think you are done, re-evaluate the project again by listing and reading the files to ensure that everything looks good. If not generate a new plan on how to fix those issues that you see.

This is a common error is encountered so make sure you generate code that does not run into this error:
npm run build

> tennis-racket-marketplace@1.0.0 build
> vue-cli-service build

⠼  Building for production...

 ERROR  Failed to compile with 1 error                                                                                                                6:39:45 AM

This relative module was not found:

* ./src/main.js in multi ./src/main.js
 ERROR  Build failed with errors.