Prompt 1:
I have an incomplete project with a jumble of 'Vue2 Component's and 'Vue2.7+ SFC's in the components/ directory of the project but everything else from a proper Vue 3 application is missing. Some of the Vue2.7+ SFCs have .js extensions but they are not .js files, treat them as .vue files. If you need to convert Vue 2 components into Vue 3 SFCs, do so. I want you to turn this project into a proper Vue 3 project for me as if I started from scratch and just added in the components/*.vue files in. The Vue 3 application make use of and display all of the vue 3 components that we have.

Transcript in tmp/PROMPT1_TRANSCRIPT.html